+++ components/animation/index.ts
@@ -6,23 +6,3 @@ export {
   useStaggerAnimation,
   useParallax,
 } from './gsap-provider';
-
-// Animation Components
-export {
-  AnimatedSection,
-  AnimatedCard,
-  AnimatedText,
-  AnimatedStagger,
-  AnimatedCounter,
-  Parallax,
-  Reveal,
-} from './animated-components';
-
-// Background & Decorative
-export {
-  AnimatedBackground,
-  HeroGlow,
-  FireGlow,
-  ScrollIndicator,
-  SectionDivider,
-} from './animated-background';