+++ components/gamification/index.ts
@@ -1,18 +1,5 @@
 // Gamification Components
 export {
-  StreakDisplay,
-  StreakBadge,
-  StreakIndicator,
-  WeeklyProgress,
-} from './streak-display';
-
-export {
-  ScoreChart,
-  ScoreSparkline,
-  ScoreTrend,
-} from './score-chart';
-
-export {
   BadgeGrid,
   BadgeShowcase,
   BadgeDisplay,